Create Question6
The function is given an array of characters. Compress the array into a string using the following rules. For each group of consecutively repeating characters:
If the number of repeating characters is one, append the string with only this character.
If the number n of repeating characters x is greater than one, append the string with "x" + n.
Example of expected output: compress (["a", "a", "b", "b", "c", "c", "c"])
➞ "a2b2c3"